# Backend/src/swagger/generated-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebEngineering Flashcards API",
    "version": "1.0.0",
    "description": "A REST API for managing flashcards and folders",
    "contact": {
      "name": "API Support",
      "email": "support@karteikarten-app.com"
    }
  },
  "servers": [
    {
      "url": "https://web-engineering-karteikarten.vercel.app",
      "description": "Production server (Vercel)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in format: Bearer <token>"
      }
    },
    "schemas": {
      "Card": {
        "type": "object",
        "required": [
          "title",
          "question",
          "answer",
          "folderId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique ID of the flashcard",
            "example": "1a2b3c4d-5e6f-47ab-8c9d-0e1f2a3b4c5a"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Title of the flashcard",
            "example": "JavaScript Closures"
          },
          "question": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "Question on the flashcard",
            "example": "What is a closure in JavaScript?"
          },
          "answer": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "Answer on the flashcard",
            "example": "A closure is a function that has access to variables from its outer scope."
          },
          "currentLearningLevel": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5,
            "description": "Current learning level (0-5)",
            "example": 2
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation date",
            "example": "2024-05-01T08:00:00.000Z"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "Tags for categorization",
            "example": [
              "javascript",
              "frontend",
              "advanced"
            ]
          },
          "folderId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the associated folder",
            "example": "c6f8fb2b-a33f-46da-941d-9832b6212395"
          }
        }
      },
      "Folder": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique ID of the folder",
            "example": "c6f8fb2b-a33f-46da-941d-9832b6212395"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Name of the folder",
            "example": "JavaScript Basics"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID of the parent folder",
            "example": "parent-folder-uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation date",
            "example": "2024-05-01T08:00:00.000Z"
          },
          "lastOpenedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last opened date",
            "example": "2024-05-15T14:30:00.000Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "fail",
              "error"
            ],
            "description": "Error status (fail = client error, error = server error)",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Something went wrong"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/cards": {
      "get": {
        "summary": "Get all cards with advanced filtering and sorting options",
        "description": "Search all cards system-wide with various filtering options",
        "tags": [
          "Cards"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "folderId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by specific folder"
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "string"
            },
            "description": "Filter by tags (comma-separated)",
            "example": "javascript,frontend"
          },
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "Filter by title (partial match)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of cards per page"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of cards to skip for pagination"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "currentLearningLevel"
              ]
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "List of filtered cards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cards": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Card"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "limit": {
                              "type": "integer"
                            },
                            "offset": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new card",
        "description": "Creates a new flashcard and assigns it to a folder",
        "tags": [
          "Cards"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "question",
                  "answer",
                  "folderId"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Card title",
                    "example": "JavaScript Closures"
                  },
                  "question": {
                    "type": "string",
                    "description": "Question on the front side",
                    "example": "What is a closure in JavaScript?"
                  },
                  "answer": {
                    "type": "string",
                    "description": "Answer on the back side",
                    "example": "A closure is a function that has access to variables from its outer scope..."
                  },
                  "folderId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the associated folder"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Tags for categorization",
                    "example": [
                      "javascript",
                      "closures",
                      "functions"
                    ]
                  },
                  "currentLearningLevel": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "default": 1,
                    "description": "Current learning level (1-5)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Card successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "card": {
                          "$ref": "#/components/schemas/Card"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Folder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/folders": {
      "get": {
        "summary": "Get all folders with pagination",
        "description": "Loads all available folders with optional pagination",
        "tags": [
          "Folders"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of folders per page"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of folders to skip for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all folders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "folders": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Folder"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "limit": {
                              "type": "integer"
                            },
                            "offset": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new folder",
        "description": "Creates a new folder (can also be a subfolder)",
        "tags": [
          "Folders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Folder name",
                    "example": "JavaScript Grundlagen"
                  },
                  "parentId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the parent folder (optional for subfolders)",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "folder": {
                          "$ref": "#/components/schemas/Folder"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Übergeordneter Folder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}